requirements.txt
streamlit==1.31.1
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
python-dotenv==1.0.1
dspy-ai==0.1.9
